credi_rag/requirements.txt
pandas>=2.2.2
numpy>=1.26.4
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
transformers>=4.42.0
torch>=2.2.2
streamlit>=1.36.0
openai>=1.40.0
tabulate>=0.9.0
